refactor: consolidate qiskit_fermions.protocols
