+++ docs/pydoc/qiskit_fermions.protocols.rst
@@ -1,0 +1,4 @@
+.. automodule:: qiskit_fermions.protocols
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
