+++ python/qiskit_fermions/_lib/operators/operators_terms/ordering/canonical/__init__.pyi
@@ -1,7 +1,7 @@
 # This file is automatically generated by pyo3_stub_gen
 # ruff: noqa: E501, F401, F403, F405
 
-from qiskit_fermions.operators.protocol import OperatorTrait
+from qiskit_fermions.operators.operator_trait import OperatorTrait
 __all__ = [
     "canonical_order",
 ]
