+++ python/qiskit_fermions/linalg/__init__.py
@@ -27,6 +27,7 @@ This module provides various linear algebra utilities.
    givens_decomposition_slater
    double_factorized_2body
    double_factorized_t2
+   linear_operator
    reconstruct_t2
    double_factorized_t2_alpha_beta
    reconstruct_t2_alpha_beta
@@ -44,12 +45,15 @@ from qiskit_fermions._lib.linalg.givens import (
     givens_decomposition_slater,
 )
 
+from .linear_operator import linear_operator
+
 __all__ = [
     "double_factorized_2body",
     "double_factorized_t2",
     "double_factorized_t2_alpha_beta",
     "givens_decomposition",
     "givens_decomposition_slater",
+    "linear_operator",
     "reconstruct_t2",
     "reconstruct_t2_alpha_beta",
 ]
